Fixing C4Slider
There were problems with copying sliders… Figured out a new way of
copying styles between UISliders by simply passing the slider from the
old object as an entry in the style dictionary to be copied. Then,
change the values of the new slider based on the copy of the old slider.