Tweaking C4Stepper
Realized that it's not possible to copy the images from inside a
UISlider and reuse them. This is because the stepper creates a
stretched image from the original one you pass it, and it's not
possible to retrieve the original image. So, I've eliminated passing
images via the C4Stepper style.