Adding functionality to C4Control
Can now toggle user interaction of its view, and can also remove the
control from any other view.

NOTE: this might not be an elegant solution, because we are now
clashing between view / control behaviour a bit. But, to be able to add
a control, without being able to remove it is not proper either.

Also, to add the method removeFromSupercontrol is also an awkward clash.