Checks anchorPoint prior to setting endPoints for origin/center
This fix checks to see if the anchor point of the receiver has its
default value. If not, this implies that the developer has changed the
position of the anchorPoint. In such a situation it is not easy to
determine where the endPoints should actually go), so we default to
changing the center of the view rather than updating the endPoints.